refactor: layout and loading
