--- src/components/Loading/Loading.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import './Loading.scss'
-
-class Loading extends React.Component {
-  render() {
-    return <h1 className="loading block palette-invert">Loading</h1>
-  }
-}
-
-export default Loading
