--- src/components/Loading/Loading.scss
@@ -1,6 +1,0 @@
-.loading {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
