--- src/components/Loading/index.js
@@ -1,2 +1,0 @@
-import Loading from './Loading'
-export default Loading
